feat: cambia tipo de dato peso a float (bags)

+++ colanta-backend/App/Bags/Bag.cs
@@ -1,6 +1,6 @@
 namespace colanta_backend.App.Bags {
     public class Bag {
         public int VtexId { get; set; }
-        public int Weigth { get; set; } 
+        public float Weigth { get; set; } 
     }
 }
